finally able to debug using the tester

+++ test/test.ps1
@@ -28,6 +28,30 @@ TraceInfo $license
 TraceInfo "Start to set initialize data disk(s)"
 Get-Disk | Where partitionstyle -eq 'raw' | Initialize-Disk -PartitionStyle MBR -PassThru |	New-Partition -AssignDriveLetter -UseMaximumSize | Format-Volume -FileSystem NTFS -Confirm:$false
 TraceInfo "Finished formatting data disk(s)"
+
+
+TraceInfo "Installing DataKeeper license"
+$licfile = $env:windir\SysWOW64\LKLicense\extmirrsvc.lic
+Add-Content $licfile $license
+TraceInfo "Finished installing DataKeeper license"
+
+TraceInfo "Enabling and Configuring WSFC"
+Install-WindowsFeature -Name Failover-Clustering -IncludeManagementTools
+Import-Module FailoverClusters
+# TODO: add node to cluster 
+TraceInfo "Finished Configuring WSFC"
+
+TraceInfo "Verifying ExtMirr Service is Running"
+while(-Not $(Test-Path $env:windir\SysWOW64\LKLicense\extmirrsvc.lic)) {
+		Start-Sleep -Seconds 5
+}
+Restart-Service extmirrsvc
+Start-Sleep -Seconds 10
+if(-Not $(Get-Service extmirrsvc).Status -eq "Running") {
+	TraceInfo "ExtMirr Service failed to start. Mirror not created."
+	return 1
+}
+
 	
 TraceInfo "Restarting this node after 30 seconds"
 Start-Process -FilePath "cmd.exe" -ArgumentList "/c shutdown /r /t 30"
